bugfix -- memory alloc

+++ HiddenMarkovModel.h
@@ -37,7 +37,7 @@ private:
     StochasticMatrix transitionMatrix;
     StochasticMatrix observationMatrix;
     StochasticRow initialState;
-    std::vector<int> scalingFactors;
+    std::vector<double> scalingFactors;
 
     Matrix alphaPass(const ObservationSequence& O);
     Matrix betaPass(const ObservationSequence& O);
